whishlist updated

+++ BabyBackend/Services/WhishListService/WhishListServices.cs
@@ -1,10 +1,13 @@
 ﻿using AutoMapper;
 using BabyBackend.Controllers;
 using BabyBackend.DbContexts;
+using BabyBackend.JwtVerification;
 using BabyBackend.Migrations;
 using BabyBackend.Models;
 using BabyBackend.Models.Dto;
 using Microsoft.EntityFrameworkCore;
+using Newtonsoft.Json.Linq;
+using Razorpay.Api;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -17,19 +20,26 @@ namespace BabyBackend.Services.WhishListService
         private readonly IMapper _mapper;
         private readonly IConfiguration _configuration;
         private readonly string HostUrl;
+        private readonly IJwtServices _jwtServices;
 
 
-        public WhishListServices(BabyDbContext dbContext, IMapper mapper, IConfiguration configuration)
+        public WhishListServices(BabyDbContext dbContext, IMapper mapper, IConfiguration configuration, IJwtServices jwtServices)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _configuration = configuration;
             HostUrl = _configuration["HostUrl:url"];
+            _jwtServices = jwtServices;
 
         }
 
-        public async Task<bool> AddToWhishList(int userId, int productId)
+        public async Task<bool> AddToWhishList(string token, int productId)
         {
+            int userId = _jwtServices.GetUserIdFromToken(token);
+            if (userId == null)
+            {
+                throw new Exception("user id not valid");
+            }
             var wExist = _dbContext.whishLists.Include(w=>w.products).FirstOrDefault(w => w.ProductId == productId);
             if(wExist == null)
             {
@@ -49,9 +59,14 @@ namespace BabyBackend.Services.WhishListService
             
         }
 
-        public async Task RemoveWhishList(int productId)
+        public async Task RemoveWhishList(string token,int productId)
         {
-            var wList = await _dbContext.whishLists.FirstOrDefaultAsync(p=>p.ProductId==productId);
+            int userId = _jwtServices.GetUserIdFromToken(token);
+            if (userId == null)
+            {
+                throw new Exception("user id not valid");
+            }
+            var wList = await _dbContext.whishLists.FirstOrDefaultAsync(p=>p.UserId==userId && p.ProductId == productId);
             if (wList != null)
             {
                 _dbContext.whishLists.Remove(wList);
@@ -59,8 +74,13 @@ namespace BabyBackend.Services.WhishListService
             }
         }
 
-        public async Task<List<WhishListViewDto>> GetWhishLists(int userId)
+        public async Task<List<WhishListViewDto>> GetWhishLists(string token)
         {
+            int userId = _jwtServices.GetUserIdFromToken(token);
+            if (userId == null)
+            {
+                throw new Exception("user id not valid");
+            }
             var wList = await _dbContext.whishLists
                 .Include(w => w.products)
                 .ThenInclude(p => p.Category)
@@ -85,6 +105,22 @@ namespace BabyBackend.Services.WhishListService
                 return new List<WhishListViewDto>();
             }
             
+        }
+
+        public async Task<bool> isWishListExist(string token, int productId)
+        {
+            int userId = _jwtServices.GetUserIdFromToken(token);
+            if (userId == null)
+            {
+                throw new Exception("user id not valid");
+            }
+            var wList = await _dbContext.whishLists.FirstOrDefaultAsync(p => p.UserId == userId && p.ProductId == productId);
+            if(wList == null)
+            {
+                return true;
+            }
+            return false;
+
         }
     }
 }
